Fix invalid hashtag notation and allow substring reference

+++ src/main/java/seedu/address/logic/commands/DeleteCommand.java
@@ -8,6 +8,7 @@ import seedu.address.commons.util.ToStringBuilder;
 import seedu.address.logic.Messages;
 import seedu.address.logic.commands.exceptions.CommandException;
 import seedu.address.model.Model;
+import seedu.address.model.coursemate.ContainsKeywordPredicate;
 import seedu.address.model.coursemate.CourseMate;
 import seedu.address.model.coursemate.QueryableCourseMate;
 import seedu.address.model.coursemate.exceptions.CourseMateNotFoundException;
@@ -42,15 +43,27 @@ public class DeleteCommand extends Command {
             throw new CommandException(Messages.MESSAGE_INVALID_COURSE_MATE_DISPLAYED_INDEX);
         }
 
+        List<CourseMate> courseMateToDeleteList;
         try {
-            CourseMate courseMateToDelete = model.findCourseMate(queryableCourseMate);
-            model.deleteCourseMate(courseMateToDelete);
-            model.setRecentlyProcessedCourseMate(courseMateToDelete);
-            return new CommandResult(MESSAGE_DELETE_COURSE_MATE_SUCCESS, false, false, true);
+            courseMateToDeleteList = model.findCourseMate(queryableCourseMate);
         } catch (CourseMateNotFoundException e) {
             throw new CommandException(Messages.MESSAGE_INVALID_COURSE_MATE_NAME);
         }
 
+        //If there are more than 1 matching names
+        if (courseMateToDeleteList.size() > 1) {
+            ContainsKeywordPredicate predicate = new ContainsKeywordPredicate(
+                    queryableCourseMate.getName().toString());
+            model.updateFilteredCourseMateList(predicate);
+            return new CommandResult(
+                    String.format(Messages.MESSAGE_SIMILAR_COURSE_MATE_NAME,
+                            model.getFilteredCourseMateList().size()), false , false, true);
+        }
+
+        CourseMate courseMateToDelete = courseMateToDeleteList.get(0);
+        model.deleteCourseMate(courseMateToDelete);
+        model.setRecentlyProcessedCourseMate(courseMateToDelete);
+        return new CommandResult(MESSAGE_DELETE_COURSE_MATE_SUCCESS, false, false, true);
     }
 
     @Override
